fix(xv6-mbc): first build attempt iteration 1 — zicsr+zmmul extensions
ASCEND-LINUX Phase 1.1. First end-to-end build attempt surfaced 4
concrete issues; this commit fixes 3, parks 1 for next shift.

Fixed in this iteration:
- AS = $(TOOLPREFIX)as (was 'gas' — wrong binary name on WEST).
- ASFLAGS gain -march=rv32i_zicsr_zmmul (was rv32i; xv6 entry.S uses
  csrr + mul — need zicsr extension for CSR access mnemonics, zmmul
  for multiply).
- CFLAGS also gain rv32i_zicsr_zmmul (kernel/riscv.h has inline asm
  csrr/csrw via macros included from C source).
- CFLAGS gain `-isystem $(GCC_INCLUDE)` for stdarg.h (xv6's printf.c
  needs it; -nostdinc excluded GCC's freestanding headers; matches
  demos/doom/Makefile pattern).
- start_mbc.c: replaced (uintptr_t)main → (uint64)main since xv6's
  types.h has uint64 but no uintptr_t; on RV32 with -mabi=ilp32 both
  resolve to 32-bit.

Build progresses 7 .o files compiled successfully:
- kernel/entry.o
- kernel/console.o
- kernel/printf.o
- kernel/kalloc.o
- adapters/start_mbc.o
- adapters/console-mmio.o
- adapters/syscall_shims.o

Build now stops at kernel/swtch.S — uses RV64 `sd`/`ld` doubleword
load/store instructions. RV32 has only `sw`/`lw` (word). This
context-switch routine needs an RV32 rewrite.

🎖️ Marshal handoff for next shift: invoke /unheaded-marshal with:
  "ASCEND-LINUX P1.1 first build attempt: 7 .o files compiled green;
  blocked on kernel/swtch.S (RV64 sd/ld instructions). NEXT: replace
  upstream/kernel/swtch.S with adapters/swtch_mbc.S using sw/lw
  (32-bit). xv6's struct context (kernel/proc.h) is 14 × u64 ;
  rewrite as 14 × u32 since RV32 saved-register file is 32-bit.
  Then re-run `make -f ../adapters/Makefile.mbc kernel` and continue
  iterating. Owner Developer + Computermancer."

+++ crates/xv6-mbc/adapters/start_mbc.c
@@ -134,7 +134,9 @@ start(void)
     // MEPC = main(); the address translator places main() somewhere in the
     // MBC text region, and we encode the byte address here. The compiled MBC
     // MRET (0x47) reads MEPC>>2 and sets PC to that word index.
-    CSR_REG(CSR_MEPC) = (uint32)(uintptr_t)main;
+    // (uint32) cast directly — main is a function pointer; in RV32 it's already
+    // 32-bit so we don't need uintptr_t.
+    CSR_REG(CSR_MEPC) = (uint32)(uint64)main;
 
     // Step 4: MRET. The MBC compiler emits opcode 0x47 here.
     // (We can't use `mret` — that's RISC-V machine code. Translator handles it.)
